+++ src/components/common/LicenseInfo.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export default function LicenseInfo() {
+  return (
+    <div className="flex flex-col items-center mt-[30%] text-gray-500 text-[10px]">
+      <p>©2024 MYFASHION COORDINATOR</p>
+      <p>ALL RIGHTS RESERVED</p>
+    </div>
+  )
+}
