+++ src/components/pages/auth/SNSSignInButton.tsx
@@ -1,0 +1,9 @@
+import KakaoSignInButton from './signIn/KakaoSignInButton'
+
+export default function SNSSignInButton() {
+  return (
+    <div className="my-5">
+      <KakaoSignInButton />
+    </div>
+  )
+}
